gxm: split public SceGxm types and internal+opaque SceGxm types
