+++ vita3k/renderer/src/color_format.cpp
@@ -15,6 +15,7 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
+#include <cstddef>
 #include <gxm/types.h>
 
 namespace renderer::color {
